requirements.txt
# Brain Tumor Detection System – Python Dependencies
# Install with:  pip install -r requirements.txt

# ── Deep Learning ──────────────────────────────────────────
torch==2.2.2            # PyTorch (CPU build; no CUDA required)
torchvision==0.17.2     # Dataset utilities and pretrained models

# ── Image Processing ───────────────────────────────────────
Pillow==10.3.0          # PIL image I/O
opencv-python-headless==4.9.0.80   # OpenCV without GUI dependencies
numpy==1.26.4           # Numerical arrays

# ── Web Application ────────────────────────────────────────
streamlit==1.33.0       # Web UI framework

# ── Utilities ──────────────────────────────────────────────
matplotlib==3.8.4       # Optional: plot training curves
tqdm==4.66.2            # Progress bars during training
